+++ ArcadiaDevs.Viora.Platform/Agronomic/Application/Internal/CommandServices/AgronomicStatisticIngestionService.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using ArcadiaDevs.Viora.Platform.Agronomic.Application.CommandServices;
 using ArcadiaDevs.Viora.Platform.Agronomic.Application.Internal.OutboundServices;
 using ArcadiaDevs.Viora.Platform.Agronomic.Domain.Model.Aggregates;
 using ArcadiaDevs.Viora.Platform.Agronomic.Domain.Model.Commands;
@@ -14,7 +13,7 @@ using ArcadiaDevs.Viora.Platform.Agronomic.Infrastructure.ExternalServices;
 using ArcadiaDevs.Viora.Platform.Shared.Application.Model;
 using ArcadiaDevs.Viora.Platform.Shared.Domain.Model;
 
-namespace ArcadiaDevs.Viora.Platform.Agronomic.Application.CommandServices;
+namespace ArcadiaDevs.Viora.Platform.Agronomic.Application.Internal.CommandServices;
 
 /// <summary>
 ///     Command service that ingests real agronomic statistic snapshots.