+++ ArcadiaDevs.Viora.Platform/Agronomic/Interfaces/Rest/Controllers/DynamicNutritionPlansController.cs
@@ -4,6 +4,7 @@ using ArcadiaDevs.Viora.Platform.Agronomic.Domain.Model.Commands;
 using ArcadiaDevs.Viora.Platform.Agronomic.Domain.Model.Queries;
 using ArcadiaDevs.Viora.Platform.Agronomic.Interfaces.Rest.Resources;
 using ArcadiaDevs.Viora.Platform.Agronomic.Interfaces.Rest.Transform;
+using ArcadiaDevs.Viora.Platform.Iam.Infrastructure.Pipeline.Middleware.Attributes;
 using ArcadiaDevs.Viora.Platform.Shared.Domain.Model;
 using ArcadiaDevs.Viora.Platform.Shared.Resources.Errors;
 using Microsoft.AspNetCore.Http;
@@ -18,6 +19,7 @@ namespace ArcadiaDevs.Viora.Platform.Agronomic.Interfaces.Rest.Controllers;
 /// </summary>
 [ApiController]
 [Route("api/v1/dynamic-nutrition-plans")]
+[Authorize]
 public class DynamicNutritionPlansController(
     IRecommendDynamicNutritionPlanCommandService recommendDynamicNutritionPlanCommandService,
     ICertifyNutritionApplicationCommandService certifyNutritionApplicationCommandService,