+++ ArcadiaDevs.Viora.Platform/Surveillance/Interfaces/Rest/Controllers/CommunityRiskController.cs
@@ -1,4 +1,5 @@
 using System.Net.Mime;
+using ArcadiaDevs.Viora.Platform.Iam.Infrastructure.Pipeline.Middleware.Attributes;
 using ArcadiaDevs.Viora.Platform.Surveillance.Application.QueryServices;
 using ArcadiaDevs.Viora.Platform.Surveillance.Domain.Model.Queries;
 using ArcadiaDevs.Viora.Platform.Surveillance.Interfaces.Rest.Resources;
@@ -12,6 +13,7 @@ namespace ArcadiaDevs.Viora.Platform.Surveillance.Interfaces.Rest.Controllers;
 [ApiController]
 [Route("api/v1/community-risk")]
 [Produces(MediaTypeNames.Application.Json)]
+[Authorize]
 public class CommunityRiskController(ICommunityRiskQueryService communityRiskQueryService) : ControllerBase
 {
     /// <summary>