+++ ArcadiaDevs.Viora.Platform/Surveillance/Interfaces/Rest/Controllers/PestSightingReportsController.cs
@@ -1,4 +1,5 @@
 using System.Net.Mime;
+using ArcadiaDevs.Viora.Platform.Iam.Infrastructure.Pipeline.Middleware.Attributes;
 using ArcadiaDevs.Viora.Platform.Shared.Resources.Errors;
 using ArcadiaDevs.Viora.Platform.Surveillance.Application.CommandServices;
 using ArcadiaDevs.Viora.Platform.Surveillance.Interfaces.Rest.Resources;
@@ -12,6 +13,7 @@ namespace ArcadiaDevs.Viora.Platform.Surveillance.Interfaces.Rest.Controllers;
 [ApiController]
 [Route("api/v1/pest-sighting-reports")]
 [Produces(MediaTypeNames.Application.Json)]
+[Authorize]
 public class PestSightingReportsController(
     IPestSightingCommandService commandService,
     IStringLocalizer<ErrorMessages> errorLocalizer,